+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed missing keywords and categories in `Cargo.toml`.
 - Fixed missing `Error::IsTerminal` error message.
+- Fixed file handling (error is returned when file is terminal).
 
 ### Changed